// my-angular-app/src/app/models/viewModelFixtures.ts
export interface ViewModelFixture {
    fixture_id: number,
    event_date: number;
    homeTeam_id: number;
    homeTeam_name: string;
    homeTeam_logo: string;
    goalsHomeTeam: number;
    awayTeam_id: number;
    awayTeam_name: string;
    awayTeam_logo: string;
    goalsAwayTeam: number;
    statusShort: string;
  }